refactor: split frontend app state and page flows
