+++ front-end/src/app/App.tsx
@@ -8,6 +8,7 @@ import { ItemDetailPage } from "./components/ItemDetailPage";
 import { MyOutfitsPage } from "./components/MyOutfitsPage";
 import { UserDetailPage } from "./components/UserDetailPage";
 import { UsersDirectoryPage } from "./components/UsersDirectoryPage";
+import { AccessRestrictedState } from "./components/shared/AccessRestrictedState";
 import { Input } from "./components/ui/input";
 import {
   Select,
@@ -19,156 +20,35 @@ import {
 import {
   beginGoogleSignIn,
   ClothingItem,
-  CreateItemMode,
-  emptyOutfitDraft,
-  fetchClosetOwner,
+  fetchCurrentUser,
   formatPossessive,
   formatPreferredStyle,
-  loadOutfitDraft,
   logoutSession,
   OutfitDraft,
-  saveOutfitDraft,
   titleize,
   User,
 } from "./lib/closet";
-
-interface RouteState {
-  kind: "home";
-}
-
-interface ClosetRouteState {
-  kind: "closet";
-}
-
-interface ItemRouteState {
-  kind: "item";
-  itemId: number;
-}
-
-interface UsersRouteState {
-  kind: "users";
-}
-
-interface UserRouteState {
-  kind: "user";
-  userId: number;
-}
-
-interface NewItemRouteState {
-  kind: "new-item";
-  userId: number | null;
-  mode: CreateItemMode;
-}
-
-interface OutfitsRouteState {
-  kind: "outfits";
-}
-
-interface NotFoundRouteState {
-  kind: "not-found";
-}
+import {
+  buildTagOptions,
+  ClosetSortOption,
+  filterClothingItems,
+  hasActiveClosetControls,
+} from "./lib/closetFilters";
+import {
+  AppRoute,
+  authErrorMessage,
+  getRouteFromLocation,
+  isClosetRoute,
+  isOutfitRoute,
+  isProtectedRoute,
+  isUsersRoute,
+  navigateTo,
+} from "./lib/routes";
+import { useOutfitDraftState } from "./lib/useOutfitDraftState";
 
 interface HomeMessageState {
   kind: "error" | "success";
   text: string;
-}
-
-type AppRoute =
-  | RouteState
-  | ClosetRouteState
-  | ItemRouteState
-  | UsersRouteState
-  | UserRouteState
-  | NewItemRouteState
-  | OutfitsRouteState
-  | NotFoundRouteState;
-
-function isClosetRoute(route: AppRoute) {
-  return route.kind === "closet" || route.kind === "item" || route.kind === "new-item";
-}
-
-function isOutfitRoute(route: AppRoute) {
-  return route.kind === "outfits";
-}
-
-function isUsersRoute(route: AppRoute) {
-  return route.kind === "users" || route.kind === "user";
-}
-
-function isProtectedRoute(route: AppRoute) {
-  return route.kind !== "home" && route.kind !== "not-found";
-}
-
-function authErrorMessage(code: string | null) {
-  switch (code) {
-    case "auth_cancelled":
-      return "Google sign-in was cancelled before it finished.";
-    case "google_auth_failed":
-      return "Google sign-in could not be completed. Please try again.";
-    case "signin_failed":
-      return "Sign-in failed. Please try again.";
-    default:
-      return "We couldn't sign you in. Please try again.";
-  }
-}
-
-function parseCreateItemMode(value: string | null): CreateItemMode {
-  return value === "image" ? "image" : "manual";
-}
-
-function getRouteFromLocation(
-  pathname = window.location.pathname,
-  search = window.location.search,
-): AppRoute {
-  const normalizedPath = pathname.replace(/\/+$/, "") || "/";
-  const itemMatch = normalizedPath.match(/^\/items\/(\d+)$/);
-  const userMatch = normalizedPath.match(/^\/users\/(\d+)$/);
-  const query = new URLSearchParams(search);
-
-  if (normalizedPath === "/items/new") {
-    const userId = query.get("userId");
-    return {
-      kind: "new-item",
-      userId: userId ? Number(userId) : null,
-      mode: parseCreateItemMode(query.get("mode")),
-    };
-  }
-
-  if (normalizedPath === "/closet") {
-    return { kind: "closet" };
-  }
-
-  if (normalizedPath === "/users") {
-    return { kind: "users" };
-  }
-
-  if (normalizedPath === "/outfits") {
-    return { kind: "outfits" };
-  }
-
-  if (userMatch) {
-    return { kind: "user", userId: Number(userMatch[1]) };
-  }
-
-  if (itemMatch) {
-    return { kind: "item", itemId: Number(itemMatch[1]) };
-  }
-
-  if (normalizedPath === "/") {
-    return { kind: "home" };
-  }
-
-  return { kind: "not-found" };
-}
-
-function navigateTo(pathname: string) {
-  const nextUrl = new URL(pathname, window.location.origin);
-  if (window.location.pathname === nextUrl.pathname && window.location.search === nextUrl.search) {
-    return;
-  }
-
-  window.history.pushState({}, "", `${nextUrl.pathname}${nextUrl.search}`);
-  window.dispatchEvent(new PopStateEvent("popstate"));
 }
 
 function updateUserItem(user: User | null, nextItem: ClothingItem) {
@@ -195,50 +75,6 @@ function removeUserItem(user: User | null, itemId: number) {
   };
 }
 
-type ClosetSortOption = "name-asc" | "newest-added" | "oldest-added" | "recent-purchase";
-
-function matchesSearchQuery(item: ClothingItem, query: string) {
-  const normalizedQuery = query.trim().toLowerCase();
-  if (!normalizedQuery) {
-    return true;
-  }
-
-  const haystack = [
-    item.name,
-    item.size,
-    ...item.tags,
-  ]
-    .filter(Boolean)
-    .join(" ")
-    .toLowerCase();
-
-  return normalizedQuery
-    .split(/\s+/)
-    .every((term) => haystack.includes(term));
-}
-
-function sortClothingItems(items: ClothingItem[], sortOption: ClosetSortOption) {
-  const sorted = [...items];
-
-  sorted.sort((left, right) => {
-    if (sortOption === "name-asc") {
-      return left.name.localeCompare(right.name);
-    }
-
-    if (sortOption === "oldest-added") {
-      return (new Date(left.created_at ?? 0).getTime()) - (new Date(right.created_at ?? 0).getTime());
-    }
-
-    if (sortOption === "recent-purchase") {
-      return (new Date(right.date ?? 0).getTime()) - (new Date(left.date ?? 0).getTime());
-    }
-
-    return (new Date(right.created_at ?? 0).getTime()) - (new Date(left.created_at ?? 0).getTime());
-  });
-
-  return sorted;
-}
-
 export default function App() {
   const [route, setRoute] = useState<AppRoute>(() => getRouteFromLocation());
   const [user, setUser] = useState<User | null>(null);
@@ -248,8 +84,8 @@ export default function App() {
   const [searchQuery, setSearchQuery] = useState("");
   const [selectedTag, setSelectedTag] = useState("all");
   const [sortOption, setSortOption] = useState<ClosetSortOption>("name-asc");
-  const [outfitDraft, setOutfitDraft] = useState<OutfitDraft>(emptyOutfitDraft());
   const [outfitDraftNotice, setOutfitDraftNotice] = useState("");
+  const [outfitDraft, setOutfitDraft] = useOutfitDraftState(user);
 
   useEffect(() => {
     const handlePopState = () => setRoute(getRouteFromLocation());
@@ -281,12 +117,12 @@ export default function App() {
     const controller = new AbortController();
     const unauthorizedMessage = "You do not have permission to view this page. Please log in.";
 
-    async function loadProtectedRoute() {
+    void (async () => {
       setIsLoading(true);
       setErrorMessage("");
 
       try {
-        const nextUser = await fetchClosetOwner(controller.signal);
+        const nextUser = await fetchCurrentUser(controller.signal);
         if (!nextUser) {
           setUser(null);
           setHomeMessage({ kind: "error", text: unauthorizedMessage });
@@ -308,34 +144,10 @@ export default function App() {
           setIsLoading(false);
         }
       }
-    }
-
-    loadProtectedRoute();
+    })();
 
     return () => controller.abort();
   }, [route.kind]);
-
-  useEffect(() => {
-    if (!user) {
-      setOutfitDraft(emptyOutfitDraft());
-      return;
-    }
-
-    const availableItemIds = new Set(user.clothing_items.map((item) => item.id));
-    const persisted = loadOutfitDraft(user.id);
-    setOutfitDraft({
-      ...persisted,
-      itemIds: persisted.itemIds.filter((itemId) => availableItemIds.has(itemId)),
-    });
-  }, [user?.id, user?.clothing_items]);
-
-  useEffect(() => {
-    if (!user) {
-      return;
-    }
-
-    saveOutfitDraft(user.id, outfitDraft);
-  }, [outfitDraft, user]);
 
   useEffect(() => {
     if (!outfitDraftNotice) {
@@ -352,13 +164,22 @@ export default function App() {
   const clothingItems = user?.clothing_items ?? [];
   const deferredSearchQuery = useDeferredValue(searchQuery);
   const isLoggedOutProtectedRoute = isProtectedRoute(route) && !user;
-
   const closetTitle = user ? formatPossessive(titleize(user.username)) : "Your Closet";
   const preferredStyle = formatPreferredStyle(user?.preferred_style);
   const selectedItem =
     route.kind === "item"
       ? clothingItems.find((item) => item.id === route.itemId) ?? null
       : null;
+  const isAdminRoute = route.kind === "users" || route.kind === "user";
+  const isUnauthorizedAdminRoute = Boolean(user && !user.admin && isAdminRoute);
+  const tagOptions = buildTagOptions(clothingItems);
+  const filteredClothingItems = filterClothingItems(
+    clothingItems,
+    deferredSearchQuery,
+    selectedTag,
+    sortOption,
+  );
+  const hasActiveFilters = hasActiveClosetControls(searchQuery, selectedTag, sortOption);
 
   function addItemToOutfitDraft(itemId: number) {
     setOutfitDraft((current) => {
@@ -375,38 +196,22 @@ export default function App() {
     });
   }
 
-  const isAdminRoute = route.kind === "users" || route.kind === "user";
-  const isUnauthorizedAdminRoute = Boolean(user && !user.admin && isAdminRoute);
-  const tagOptions = Array.from(new Set(clothingItems.flatMap((item) => item.tags))).sort((left, right) =>
-    left.localeCompare(right),
-  );
-  const filteredClothingItems = sortClothingItems(
-    clothingItems.filter((item) => {
-      if (selectedTag !== "all" && !item.tags.includes(selectedTag)) {
-        return false;
-      }
+  async function handleLogout() {
+    try {
+      await logoutSession();
+    } catch (error) {
+      setErrorMessage(error instanceof Error ? error.message : "Unable to sign out right now.");
+      return;
+    }
 
-      return matchesSearchQuery(item, deferredSearchQuery);
-    }),
-    sortOption,
-  );
-  const hasActiveClosetControls =
-    searchQuery.trim().length > 0 || selectedTag !== "all" || sortOption !== "name-asc";
+    setUser(null);
+    setHomeMessage({ kind: "success", text: "Signed out successfully." });
+    navigateTo("/");
+  }
 
   const globalAction = user ? (
     <button
-      onClick={async () => {
-        try {
-          await logoutSession();
-        } catch (error) {
-          setErrorMessage(error instanceof Error ? error.message : "Unable to sign out right now.");
-          return;
-        }
-
-        setUser(null);
-        setHomeMessage({ kind: "success", text: "Signed out successfully." });
-        navigateTo("/");
-      }}
+      onClick={() => void handleLogout()}
       className="inline-flex items-center justify-center gap-3 border border-border px-4 py-2.5 text-sm transition-colors hover:border-foreground"
       style={{ fontFamily: "Outfit, sans-serif" }}
     >
@@ -505,30 +310,11 @@ export default function App() {
     );
   } else if (isUnauthorizedAdminRoute) {
     pageContent = (
-      <div className="min-h-screen bg-background">
-        <div className="max-w-3xl mx-auto px-6 py-16">
-          <button
-            onClick={() => navigateTo("/closet")}
-            className="inline-flex items-center gap-2 mb-8 text-muted-foreground hover:text-foreground transition-colors"
-          >
-            <ArrowLeft className="w-4 h-4" />
-            Back
-          </button>
-
-          <div className="border border-destructive/20 bg-destructive/5 p-8">
-            <p
-              className="uppercase tracking-[0.3em] text-xs text-destructive/80 mb-3"
-              style={{ fontFamily: "Outfit, sans-serif" }}
-            >
-              Access Restricted
-            </p>
-            <h1 className="mb-3">You are not authorized to view this page.</h1>
-            <p className="text-muted-foreground" style={{ fontFamily: "Outfit, sans-serif" }}>
-              You&apos;re not authorized to view this page.
-            </p>
-          </div>
-        </div>
-      </div>
+      <AccessRestrictedState
+        backLabel="Back"
+        message="You're not authorized to view this page."
+        onBack={() => navigateTo("/closet")}
+      />
     );
   } else if (route.kind === "users") {
     pageContent = (
@@ -549,8 +335,7 @@ export default function App() {
       />
     );
   } else if (route.kind === "new-item") {
-    const targetUser =
-      route.userId && user?.id === route.userId ? user : route.userId ? null : user;
+    const targetUser = route.userId && user?.id === route.userId ? user : route.userId ? null : user;
     const targetUserId = route.userId ?? user?.id ?? null;
 
     pageContent = (
@@ -592,7 +377,7 @@ export default function App() {
         onItemSaved={(nextItem) => setUser((current) => updateUserItem(current, nextItem))}
         onItemDeleted={(itemId) => {
           setUser((current) => removeUserItem(current, itemId));
-          setOutfitDraft((current) => ({
+          setOutfitDraft((current: OutfitDraft) => ({
             ...current,
             itemIds: current.itemIds.filter((id) => id !== itemId),
           }));
@@ -735,7 +520,7 @@ export default function App() {
                 </Select>
               </div>
 
-              {hasActiveClosetControls ? (
+              {hasActiveFilters ? (
                 <div className="flex items-center justify-between gap-4 border border-border bg-card px-4 py-3">
                   <p className="text-sm text-muted-foreground" style={{ fontFamily: "Outfit, sans-serif" }}>
                     Refine your closet with free-text search, tag filters, and sorting.
@@ -753,7 +538,6 @@ export default function App() {
                   </button>
                 </div>
               ) : null}
-
             </motion.div>
 
             {user && filteredClothingItems.length > 0 ? (
@@ -775,7 +559,7 @@ export default function App() {
                 </p>
                 <p className="text-muted-foreground" style={{ fontFamily: "Outfit, sans-serif" }}>
                   {user
-                    ? hasActiveClosetControls
+                    ? hasActiveFilters
                       ? "Try a different tag, search phrase, or sort."
                       : "Add a new item to start building out this closet."
                     : "Sign in with Google to load your closet."}
@@ -858,7 +642,8 @@ export default function App() {
           className="fixed bottom-6 right-6 z-50 max-w-sm border border-foreground/20 bg-background/95 px-4 py-3 text-sm shadow-lg backdrop-blur"
           style={{ fontFamily: "Outfit, sans-serif" }}
         >
-          {outfitDraftNotice} Draft has {outfitDraft.itemIds.length} {outfitDraft.itemIds.length === 1 ? "item" : "items"}.
+          {outfitDraftNotice} Draft has {outfitDraft.itemIds.length}{" "}
+          {outfitDraft.itemIds.length === 1 ? "item" : "items"}.
         </motion.div>
       ) : null}
